feat: look up user profile (#6)
