+++ mypong/mypong/settings.py
@@ -42,6 +42,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'rest_framework',
     'users',
+    'user_profile',
 ]
 
 MIDDLEWARE = [
