+++ mypong/user_profile/serializers.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from users.models import User
+
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = '__all__'
