+++ mypong/user_profile/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from .views import UserDetailAPI
+
+app_name = 'user_profile'
+
+urlpatterns = [
+    path('profile/', UserDetailAPI.as_view()),
+]
