+++ mypong/user_profile/views.py
@@ -1,0 +1,14 @@
+from rest_framework import generics
+
+from users.models import User
+from .serializers import UserSerializer
+
+
+class UserCreationAPI(generics.ListCreateAPIView):
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
+
+
+class UserDetailAPI(generics.RetrieveAPIView):
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
