+++ mypong/users/serializers.py
@@ -1,9 +1,1 @@
 from rest_framework import serializers
-
-from .models import User
-
-
-class UserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = '__all__'
