+++ mypong/users/urls.py
@@ -1,7 +1,6 @@
-from django.urls import path
+from django.urls import path, include
 
 from . import views
-from .views import UserCreationAPI, UserDetailAPI
 
 app_name = 'users'
 
@@ -9,6 +8,5 @@ urlpatterns = [
     # /users/hello
     path('hello/', views.HelloAPI),
 
-    path('create/', UserCreationAPI.as_view()),
-    path('<pk>/', UserDetailAPI.as_view()),
+    path('<pk>/', include('user_profile.urls')),
 ]
